+++ .storybook/main.js
@@ -1,0 +1,4 @@
+module.exports = {
+  stories: ['../src/stories/**/*.stories.js'],
+  addons: ['@storybook/addon-actions', '@storybook/addon-links'],
+};
