+++ README.md
@@ -4,5 +4,3 @@ A normal radio set allows you to select one unique value exclusive to a set.
 This component allows you to select a unique value across multiple sets that share the same set of values. When an option is selected in one set, it is automatically disabled in associated sets.
 
 [View demo](http://dogsghost.github.io/rc-unique-multi-select)
-
-This project was bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).
