+++ package.json
@@ -2,87 +2,23 @@
   "name": "rc-unique-multi-select",
   "homepage": "http://dogsghost.github.io/rc-unique-multi-select",
   "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
+  "dependencies": {},
   "devDependencies": {
-    "autoprefixer": "6.4.1",
-    "babel-core": "6.14.0",
-    "babel-eslint": "6.1.2",
-    "babel-jest": "15.0.0",
-    "babel-loader": "6.2.5",
-    "babel-preset-react-app": "^0.2.1",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "chalk": "1.1.3",
-    "connect-history-api-fallback": "1.3.0",
-    "cross-spawn": "4.0.0",
-    "css-loader": "0.24.0",
-    "detect-port": "1.0.0",
-    "dotenv": "2.0.0",
-    "eslint": "3.5.0",
-    "eslint-config-react-app": "^0.2.1",
-    "eslint-loader": "1.5.0",
-    "eslint-plugin-flowtype": "2.18.1",
-    "eslint-plugin-import": "1.12.0",
-    "eslint-plugin-jsx-a11y": "2.2.2",
-    "eslint-plugin-react": "6.3.0",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
-    "filesize": "3.3.0",
-    "find-cache-dir": "0.1.1",
-    "fs-extra": "0.30.0",
-    "gzip-size": "3.0.0",
-    "html-webpack-plugin": "2.22.0",
-    "http-proxy-middleware": "0.17.1",
-    "jest": "15.1.1",
-    "json-loader": "0.5.4",
-    "object-assign": "4.1.0",
-    "path-exists": "2.1.0",
-    "postcss-loader": "0.13.0",
-    "promise": "7.1.1",
-    "prop-types": "^15.7.2",
-    "react-dev-utils": "^0.2.1",
-    "recursive-readdir": "2.1.0",
-    "rimraf": "2.5.4",
-    "strip-ansi": "3.0.1",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.2",
-    "webpack-dev-server": "1.16.1",
-    "whatwg-fetch": "1.0.0"
+    "@babel/core": "^7.10.1",
+    "@storybook/addon-actions": "^5.3.19",
+    "@storybook/addon-links": "^5.3.19",
+    "@storybook/addons": "^5.3.19",
+    "@storybook/react": "^5.3.19",
+    "babel-loader": "^8.1.0"
   },
-  "dependencies": {
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2"
+  "peerDpendencies": {
+    "react": "^16.13.1"
   },
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "jsx",
-      "js",
-      "json"
-    ],
-    "moduleNameMapper": {
-      "^.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/config/jest/FileStub.js",
-      "^.+\\.css$": "<rootDir>/config/jest/CSSStub.js"
-    },
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testPathIgnorePatterns": [
-      "<rootDir>/(build|docs|node_modules)/"
-    ],
-    "testEnvironment": "node",
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(js|jsx)$"
-  },
-  "babel": {
-    "presets": [
-      "react-app"
-    ]
-  },
-  "eslintConfig": {
-    "extends": "react-app"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
   }
 }
