--- src/index.css
@@ -1,5 +1,0 @@
-body {
-  margin: 0;
-  padding: 1em;
-  font-family: sans-serif;
-}
